WIP: storyPanel refactor
