+++ lib/ReactViews/Map/Panels/StoryPanel/story-panel.scss
@@ -1,0 +1,18 @@
+.storyBtn {
+  composes: btn from "../../../../Sass/common/_buttons.scss";
+  composes: btn--map from "../../../../Sass/common/_buttons.scss";
+
+  // &[aria-expanded="true"] {
+  //   background: $color-primary;
+  //   svg {
+  //     fill: #ffffff;
+  //   }
+  // }
+
+  &:hover,
+  &:focus {
+    svg {
+      fill: #ffffff;
+    }
+  }
+}
