+++ lib/ReactViews/Map/Panels/StoryPanel/story-panel.scss.d.ts
@@ -1,0 +1,7 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'storyBtn': string;
+}
+declare var cssExports: CssExports;
+export = cssExports;
