+++ lib/ReactViews/Map/menu-bar.scss
@@ -78,7 +78,7 @@
   composes: flex from "../../Sass/common/_base.scss";
 }
 
-.langPanel {
+.rightPanel {
   margin-left: $padding;
 }
