+++ lib/ReactViews/Map/menu-bar.scss.d.ts
@@ -3,7 +3,6 @@
 interface CssExports {
   'flex': string;
   'langBtn': string;
-  'langPanel': string;
   'menu': string;
   'menu-bar': string;
   'menu-bar--workbenchClosed': string;
@@ -11,6 +10,7 @@ interface CssExports {
   'menuBar': string;
   'menuBarWorkbenchClosed': string;
   'menuItem': string;
+  'rightPanel': string;
   'storyBtn': string;
 }
 declare var cssExports: CssExports;
